app: lazily install mobile toolchains
when we only had the SGX toolchain to install, things were good and
simple. but if we keep adding all toolchains to `rust-toolchain.toml`,
we'll end up with like _9_ different toolchains.

CI will need to install each toolchain each time, which will just waste
tons of CI time.

Although I think we're starting to approach the point where we need a
docker image for CI.