bgp+: Wake BGP when events need processing
Instead of having a frequent process events timer, we increase the timer
to 60s and instead implement a mpsc channel which allows various actors
notify the background processor that there are events which need to be
processed. This includes:

- Channel monitor persists (which may generate an event to resume
  monitor updating / broadcast a funding transaction)
- Opening a channel (which generates a channel open event)
- Sending a payment (which generates a payment sent event)

This commit also fixes the background processor implementation to more
closely resemble the upstream background processor implemention, i.e.
that when the future returned from
`ChannelManager::get_persistable_update_future` completes, events are
processed before the channel manager is repersisted.