test_event+: Add `TestEvent::TxBroadcasted`
Replacing `EventsProcessed` with `TxBroadcasted` fixed a subtle bug
where the condition we were *actually* interested in after a channel
open was initiated was a call to `process_events()` that occurs *after*
the `ChannelMonitorPersisted` event. Observing that the events to be
processed resulted in the resumption of channel monitor updates and the
broadcasting of a funding transaction, `TxBroadcasted` accomplishes what
we're looking for.