app: fix `FilledPlaceholder` height when replacing text
To avoid unnecessary reflows, the `FilledPlaceholder` height needs to
match whatever it's replacing.

For `Text` there is apparently an extra device-dependent
`MediaQuery.textScaleFactor` that gets multiplied in, alongside the
standard `height = fontSize * lineHeight`.

This commit adds an extra `forText` option to `FilledPlaceholder` which
will take the `MediaQuery.textScaleFactor` into account when replacing
text.