arbitrary: Remove `.no_shrink()` on txids
Turns out it actually helps to have txids which are close to 0, as it
makes it easier to visually compare what's wrong in e.g. a serialization
or ordering bug. And shrinking only happens when we find a failing test
case, so the extra computation cost isn't incurred in the standard case.