schema: Remove `sealed_seed.min_cpusvn` column
As discussed offline, it doesn't make sense to presist the `min_cpusvn`
along with each sealed seed because the minimum CPUSVN that a node
requires is already checked in git. So we remove the `min_cpusvn` from
the DB schema and will soon implement a check to compare the current
system's CPUSVN vs our `MIN_SGX_CPUSVN` checked in git.