command+: Use `send_payment_with_retry`, match on errs
The lower-level API that we tried to use just prior to this commit
(`ChannelManager::send_payment`) requires a lot of work to implement
retries. Since the fees paid could change anyway due to payment retries,
we use our computed `Route` only for fee estimation only, and let LDK
recompute the route, at the same time handling retries for us. The only
thing we have to be careful about now is our callbacks into the
`PaymentsManager`, which we implement here with strict adherence to the
LDK docs.