amount: Change internal representation to satoshis
This prevents us from having to define "display structs" to get to a
more sensible satoshi value (rather than an msat value which is too
small to be meaningful) when a field using the `Amount` newtype is
displayed using its string-serialized form, e.g. with
`serde_json::to_string_pretty`.