event_handler: Introduce `EventHandleError`
This type-enforces that if `handle_event_fallible` returns `Err`, it is
specified whether the error is tolerable or fatal, which we can then
respond to accordingly. If the error is tolerable, we simply log it and
move on; if the error is fatal, we initiate a node shutdown (coming in
subsequent commits).