bgp+event: Skip repersist if a fatal event occurs
(1) The event handler will notify the BGP (via a `fatal_event` channel)
    if event handling produced a fatal error.
(2) The BGP will skip repersisting the channel manager (as well as skip
    any peer manager I/O) if a fatal error was produced. It will then
    just shut down as there is nothing else to do.

This marks the first of two major steps required to come back into
compliance with LDK's event handler contract. See the updates to the doc
comments on `impl EventHandler` for more details.