bugfix: Don't set `reqwest::Request` timeout
Since per-request timeouts override the per-client timeout, this was
preventing `reqwest::Client`s with timeout overrides to make requests
lasting longer than `API_REQUEST_TIMEOUT`. The fix is simply to remove
this clause.

This commit also fixes one instance where a `reqwest::Client` passed in
to `from_preconfigured_client` didn't have a timeout configured.
It would be nice if we could enforce that a timeout is always set in the
`RestClient` constructors, but `reqwest` doesn't give us any API in
their `Client` or `ClientBuilder` to allow us to e.g. check that a
passed in `ClientBuilder` has a timeout set, and it's probably not worth
updating our `reqwest` fork just for this edgo case. So, we simply
document that all uses of `from_preconfigured_client` should make sure
to configure a timeout.