api: fix `PaymentIndex` not deserializing from query string correctly
It turns out that the `PaymentIndex` in e.g. `get_new_payments` was
silently getting dropped due to a `#[serde(flatten)]` + query string
limitation.

For query strings, using `#[serde(flatten)]` on a field requires that
all inner fields must be string-ish types (&str, String, Cow<'_, str>,
etc...) OR use `SerializeDisplay` and `DeserializeFromStr`.

This issue is due to a limitation in `serde`. See:
<https://github.com/serde-rs/serde/issues/1183>

The fix here is to remove `#[serde(flatten)]` for querystring
serializeable types (actually just `PaymentIndex`) and instead use
`SerializeDisplay` and `DeserializeFromStr`