bugfix: Increase `max_dust_htlc_exposure`
When we upgraded to 0.0.116, we set
`ChannelConfig::max_dust_htlc_exposure` to the default value of
`MaxDustHTLCExposure::FeeRateMultiplier(5000)`. But after tracing
backwards from the routing failure for a while I found that it was this
config was causing the `ChannelDetails::next_outbound_htlc_minimum_msat`
to end up at `2132000`. This is probably because the high priority
feerate in our integration test environment (regtest) is 237 sats/KW,
and some chain of transformations which I don't completely follow
eventually results in a minimum HTLC size of 2132 sats. Our JIT tests
happened to be using a payment size of 1000 sats, and since that's less
than `next_outbound_htlc_minimum_msat`, there would be a routing
failure deep inside `get_route`. This commit fixes the issue by
increasing our `max_dust_htlc_exposure`. Note that we are not supposed
to set this limit too high because this (supposedly) makes us vulnerable
to "fee griefing" attacks [0] which assume miner collusion.

[0]: https://lists.linuxfoundation.org/pipermail/lightning-dev/2020-May/002714.html