command+: Implement offline receive
In order to give the LSP a hook into all HTLCs routed to user nodes, we
remove all the invoice route hint logic and always include an intercept
hint, for user nodes. This allows the LSP to wake the user if a payment
comes in while the user is offline, without necessarily having to also
open a JIT channel to facilitate the payment.