minor: app-rs: Clean up `signup` API
The mock measurement shouldn't be set in the bindings and passed into
`signup`; it should be referenced directly in `signup`. The bindings is
also a weird place to put a "make an API call" TODO. This commit creates
a clear separation between the simple `signup` API and more "complex"
`signup_custom`.