node: Backup root seed during provisioning
This commit is the last in a series of changes geared towards persisting
a password-encrypted copy of the user's root seed in their Google Drive,
where the user has full control. So long as the user remembers their
password, they will be able to recover their funds in the (unlikely)
event that Lexe suddenly disappears.