refactor: Simplify `Ports` structs
We move the `user_pk` field from `UserPorts` into `RunPorts`, since
`user_pk` will eventually be only in the Run variant. We then remove
`UserPorts` and `PortReply` since these structs are now redundant.
However, much of our code still assumes that the `user_pk` field is
accessible from both variants, so we have to keep that in for now.