node::provision: Add `allow_gvfs_access`
In preparation for the implementation of revocations which require
provision instances to mutate state stored in Google Drive, we add a new
parameter `allow_gvfs_access` to `NodeProvisionRequest` which makes
explicit which provision instance (if multiple are running for the same
user at the same time) has authorization to mutate the underlying google
drive state. We also rewrite some of the provision flow.