refactor: Remove most `BackendApiError` From impls
- Also remove the `EntityConversion` `BackendErrorKind`.

The `From` impls were used in combination with ? which made everything
too implicit. It's looking like a better pattern is to define a function
like this:

```rust
impl BackendApiError {
    pub fn unauthenticated(error: impl fmt::Display) -> Self {
        let kind = BackendErrorKind::Unauthenticated;
        let msg = format!("{error:#}");
        Self { kind, msg }
    }
}
```

Then in a function which returns `Result<_, BackendApiError>` we can use
it like this:

```rust
let node_pk = signed_req
    .inner()
    .node_pk_proof
    .verify()
    .map_err(BackendApiError::unauthenticated)?;
```

It's slightly more code at the callsite but greatly enhances readability
imo.