reqwest: Remove usages of `rustls-tls-webpki-roots`
Through feature unification, it is way too easy to accidentally end up
trusting the Mozilla WebPKI root CA set when we actually intended to
trust only one root, or when using a custom TLS config. This commit
removes all usages of the `rustls-tls-webpki-roots` feature and adds a
commit to our reqwest fork so that it will fail to compile if one of the
undesired features is enabled. As a bonus, it will also prevent us from
accidentally using any of the other TLS backends available in reqwest.