app: background and burst refresh
Background Refresh

Passively resync the wallet state every ~1 min, while the user has the
wallet open. In the future, we can probably avoid this background refresh
after the first ~10 min or so with push notifications.

Burst Refresh

After send a Lightning payment, we'll want to "burst" refresh to try and
pick up the payment success quickly.