app: flutter_rust_bridge: 1.82.6 -> 2.0.0
`flutter_rust_bridge` v2.0 is a complete overhaul/rewrite and introduced
sweeping changes. My goal with this diff was to do the min. needed to
get to v2.0 while keeping everything compiling. Sadly the circular
dependency b/w the normal code and generated code makes it hard to split
up into smaller changes.

Some things are still broken, but this commit is already GIGANTIC. So
we're going to fix those up in follow-up commits.

For an overview of what changed in `flutter_rust_bridge`, see:
<https://cjycode.com/flutter_rust_bridge/#whats-new-in-v2>
<https://cjycode.com/flutter_rust_bridge/guides/miscellaneous/whats-new#primary-changes>