node: Simplify `persister::discrepancy` module
We can get all the discrepancy resolution done with just two functions.
I deduplicated the logic between channel manager and channel monitor
resolution but merged the comments so that the reasoning is still there.
These fns should be fairly easy to convert to VSS in the future.
`evaluate_and_resolve_all` will probably still be useful in the future.