app: component: simplify amount input validation
Previously, I wanted "Err(null)" in the input validator to signal
"this input is invalid, but don't bother showing an error bar/message".
Unfortunately the control flow ended up more confusing after factoring out
the `PaymentAmountInput` as a shared component. So we'll simplify the logic
here and just return `Err("")`. This shows a red error bar under the input,
which is slightly undesirable in some cases, but the control flow is simpler.