app: show payment immediately after send
Previously, we waited to `syncPayments` (which would sync the new payment into
the local DB) before navigating to the payment detail page. On local dev this
added delay is pretty much unnoticable, but the pause is jarring against
testnet.

This diff solves the issue by building a temporary "unsynced" `Payment`, using
the info available from the `pay_{onchain|invoice}` response. We then
immediately navigate to the `PaymentDetailPage` and tell it to display this
temp payment until we actually finish syncing.