bgp: Check fatal error before final persist
In the near future we might exit the `shutdown` branch at any time due
to some possibly-fatal event handling being spawned off in a task.
This is a quick workaround before we switch to LDK's new API which
should allow event handling to return a `Result`, which would let us
repersist the channel manager without losing the failed events.