event+: Handle scorer updates
Creating a different shared helper for every event type was creating too
much boilerplate, so I opted to take LDK's approach of 'pre-handling'
the event before the main event handler, although this does make the
control flow a bit non-obvious.

The main difference is we do this in our actual event handler instead of
in the BGP (as a wrapper around the event handler) like LDK does.