app-rs-codegen: use `rustup run +toolchain cargo` b/c nix `cargo` clobbers
When we enter a devShell like `app-rs-codegen` that sets the rust toolchain,
we get the actual, exact `cargo` version for that toolchain and not the
`rustup` wrapper. The actual `cargo` binary doesn't support mutli-toolchain
dispatch (ex: `cargo +nightly fmt`).

Use `rustup run` to get the older nightly `cargo fmt` so `app-rs-codegen`
works inside the nix devShell.