enclave: `debug = true` to enable panics + backtraces in enclaves
It turns out this `debug` flag is Fortanix EDP specific and is different from
the SGX platform DEBUG flag, which disables all memory protections. Instead it
just allows dumping any panics that happen inside the enclave into a host-shared
"debug buffer".

Seeing panics and their backtraces is a requirement for us to feasibly operate
nodes in production. Without them, it's almost impossible to debug any panic
that crops up.

One issue with fortanix `debug = true` is that it also preserves some registers
on abort "so that a debugger can reconstruct the stack". Ideally we would patch
this out so we don't expose any registers, but the associated build tooling
complexity required to patch rust-std makes this not worth it in the short term.