sgx: build with `panic=abort` so we get working panic+backtrace
We discovered that SGX panic machinery breaks when a panic is unwound, then
caught with `catch_unwind`, and later resumed with `resume_unwind` (e.g. all
`tokio::spawn`'ed tasks that we want to propagate panics). Instead of dumping
the panic with message+backtrace, it will just abort with `Exited with exit
code 101`, which is completely useless.

However, if we abort on panic (instead of unwinding), we get working panic
messages+backtraces.

Running with `panic=abort` isn't something we don't want long term, since it
prevents graceful shutdown on panic. In the short term though our ability to
actually debug panics in production is more important.