p2p: Actually enforce connection timeout
`p2p_handshake_timeout.is_elapsed()` would never actually return true
because the sleep future was never polled. This was an insidious error
because calling `.is_elapsed()` would count as 'using' the future, and
thus silence the lint. So I banned this `.is_elapsed()` fn to make sure
we don't shoot ourselves with this again in the future.