lexe-ln: wallet: fix BDK unable to sign when existing wallet is loaded
We were hitting test failures when the LSP would fail to open a channel, but
only when the LSP wasn't fresh, rather loaded from a previous run. The LSP
error in `FundingGenerationReady` claimed that it couldn't "sign all PSBT
inputs".

After more investigation, it turns out BDK was unable to sign _any_ TXs when
loaded from a previous state.

The fix: specify "extract_keys()" in `bdk_wallet::LoadParams`.

Why this API is FUCKING OPT IN is absolutely beyond me, but here we are...