app: Include pending LN balance in main display
If a user successfully receives funds over a zeroconf JIT channel, those
sats will not show up in the user's Lightning balance until the channel
is confirmed which is confusing. Likewise, if a User->LSP channel is
opened, the user's channel balance doesn't show up until the channel is
confirmed. This commit changes the main balance display to include
'pending' LN sats as well.