refactor: Remove `scorer_persist_[tx/rx]`
No need to notify the BGP to repersist the scorer if the scorer was
updated, as we already have the scorer on an auto-persist interval,
including a final persist at shutdown. Our scorer might get pretty big,
so we want to reduce IO; it's fine if we lose the ~5 minutes of data
caused by a crash.