api: add `LightningBalance::sendable`
Do the "sum over `next_outbound_htlc_limit`" aggregation when we build the
`NodeInfo` so the app doesn't have to fetch all channels in the normal case.

Compat: to avoid making this field `Option<Amount>`, we'll need to cut a node
release before the next app release, which will start expecting this field to
be present.