app: date format: allow for some clock drift when formatting a past DateTime
I encountered an issue where the PaymentDetailPage showed "Received * <blank>",
for a few seconds, presumably because of clock drift b/w the phone and the
node.

This diff adds some `clockDriftTolerance` to `date_format.formatDate` so we'll
consider all future DateTime's closer than `clockDriftTolerance` in duration to
be "just now".