app: receive: always show sliver of next/prev card to hint swiping
Change the `PageView.viewportFraction` calculation so that we now always show
at most 90% of the viewport. We'll switch to a fixed size page at a
`viewportWidth >= 450px` breakpoint. This way a sliver of the next card is
always visible, even on smaller screens, which helps hint to the user that
there is more content they can swipe to.

Reduce the inter-page padding s300 -> s200 to waste less screen real estate on
smaller devices.