ports: Don't reuse `ProvisionPorts` in `MegaPorts`
`ProvisionPorts` contains a useless `lexe_port` which is redundant with
the `mega_port` exposed by the mega node. We don't want to accidentally
make requests to `ProvisionPorts::lexe_port` when we meant to make the
requests to `MegaPorts::mega_port`. Also use `<consumer>_<producer>`
naming to make the APIs more clear.