client: Encode credentials with `STANDARD_NO_PAD`
We now use `STANDARD_NO_PAD` because trailing `=`s cause problems with
autocomplete on iPhone. For example, if the base64 string ends with:

- `NzB2mIn0=`
- `NzBm2In0=`

the iPhone autocompletes it to the following respectively when pasted
into iMessage, even if you 'tap away' to reject the suggestion:

- `NzB2mIn0=120 secs`
- `NzBm2In0=0 in`

`try_from_base64_blob` is now designed to accept base64 strings encoded
with either `STANDARD` or `STANDARD_NO_PAD`, as historical client
credentials were encoded with the `STANDARD` engine until we discovered
that iPhones interpret the trailing `=` as part of a unit conversion,
resulting in unintended autocompletions.