sdk-sidecar: Fix `{ payment: { ... } }` nesting
For the v2 endpoint and above, we return the response as a
[`SdkPayment`] rather than a [`SdkGetPaymentResponse`], because the
`{ "payment": { ... } }` nesting trips up dumb AIs when vibe-coding on
the Sidecar SDK, as discovered by Mat Balez et al. If the payment is
missing, we use HTTP 404 to indicate this.

If we need to add more fields to the response which don't fit in
[`SdkPayment`], we can always reintroduce the response type but with
`#[serde(flatten)]` on the [`SdkPayment`] field (since missing payments
are now indicated by HTTP 404), or do another version bump.