event+: Persist unswept outputs and txs
If we find that we cannot broadcast a `SpendableOutputs` sweep tx
because one or more of its inputs are spent or missing, we don't have
assurance that `SpendableOutputs` has been properly handled, but we also
don't want to leave the event in our event queue to be replayed over and
over again, spamming Esplora with unbroadcastable txs.

Instead, we persist the `SpendableOutputs` event and unbroadcastable tx
to separate VFS namespaces (and GDrive, if enabled) to be handled
inspected and handled later.