multi: add broadcast_hook on user node that stores txs on VFS
This commit implements a broadcast hook that store in VFS
tx information similar to previous `LspBroadcastedTx`.

Like both implementations uses the same structure to persist either
in PG or VFS, this commit also moves the definition of the struct
to `common::api::models` in order to be used in lsp, private,
backend and public.