app-rs: add OnboardingState to Settings db
In order to track the if the user has ended the backup flows, either
gdrive or seed phrase ones, we add a OnboardingState field to be
updated whenever flow is finished.

If the DB is restored, means the APP was restored in the same or
different devices. In the case there wasn't a setting.json file,
APP should still ask for backup as we don't have another way to
know if the user already backed up.